src/main/java/ru/job4j/cinema/store/SessionDBStore.java
package ru.job4j.cinema.store;

import org.apache.commons.dbcp2.BasicDataSource;
import net.jcip.annotations.ThreadSafe;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Repository;
import ru.job4j.cinema.model.Session;

import java.sql.Connection;
import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.List;
import java.util.Optional;

@ThreadSafe
@Repository
public class SessionDBStore {
    private final BasicDataSource pool;

    public SessionDBStore(BasicDataSource pool) {
        this.pool = pool;
    }

    private static final Logger LOG = LoggerFactory.getLogger(SessionDBStore.class.getName());

    public List<Session> findAll() {
        List<Session> sessions = new ArrayList<>();
        try (Connection cn = pool.getConnection();
             PreparedStatement ps = cn.prepareStatement("SELECT * FROM sessions")
        ) {
            try (ResultSet it = ps.executeQuery()) {
                while (it.next()) {
                    sessions.add(getSessionFromResultSet(it));
                }
            }
        } catch (Exception e) {
            LOG.error("exception: ", e);
        }
        return sessions;
    }

    public Session add(Session session) {
        try (Connection cn = pool.getConnection();
             PreparedStatement ps = cn.prepareStatement(
                     "INSERT INTO sessions (name,poster) VALUES (?,?)",
                     PreparedStatement.RETURN_GENERATED_KEYS)
        ) {
            ps.setString(1, session.getName());
            ps.setBytes(2, session.getPoster());
            ps.execute();
            try (ResultSet id = ps.getGeneratedKeys()) {
                if (id.next()) {
                    session.setId(id.getInt(1));
                }
            }
        } catch (Exception e) {
            LOG.error("exception: ", e);
        }
        return session;
    }

    public Optional<Session> findById(int id) {
        try (Connection cn = pool.getConnection();
             PreparedStatement ps = cn.prepareStatement("SELECT * FROM sessions WHERE id = ?")
        ) {
            ps.setInt(1, id);
            try (ResultSet it = ps.executeQuery()) {
                if (it.next()) {
                    return Optional.of(getSessionFromResultSet(it));
                }
            }
        } catch (Exception e) {
            LOG.error("exception: ", e);
        }
        return Optional.empty();
    }

    private Session getSessionFromResultSet(ResultSet it) throws SQLException {
        return new Session(it.getInt("id"),
                it.getString("name"),
                it.getBytes("poster"));
    }

}
